feat: add file type util
