+++ apps/builder/src/components/DriveFileSelect/components/FileList/index.tsx
@@ -1,5 +1,4 @@
 import { GCS_OBJECT_TYPE, IILLAFileInfo } from "@illa-public/public-types"
-import { getFileIconByContentType } from "@illa-public/utils"
 import VirtualList from "rc-virtual-list"
 import { FC, MouseEvent, useCallback, useRef } from "react"
 import { Checkbox } from "@illa-design/react"
@@ -7,6 +6,7 @@ import {
   FOLDER_LIST_CONTAINER_HEIGHT,
   FOLDER_LIST_ITEM_HEIGHT,
 } from "@/components/DriveFileSelect/constants"
+import { getFileIconByContentType } from "@/utils/file"
 import { FileListProps } from "./interface"
 import {
   fileNameStyle,
