+++ apps/builder/src/services/agent.ts
@@ -2,7 +2,7 @@ import { agentRequest, marketplaceRequest } from "@illa-public/illa-net"
 import { Agent, AgentRaw } from "@illa-public/market-agent"
 import { v4 } from "uuid"
 import { base642Blob, getFileExtensionFromBase64 } from "@/utils/file"
-import { upload } from "@/utils/file/upload"
+import { upload } from "@/utils/file"
 import { getCurrentTeamID } from "@/utils/team"
 
 export interface TeamAgentListData {
